+++ src/app/app.module.ts
@@ -3,6 +3,7 @@ import { NgModule } from '@angular/core';
 
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
+import { FirstModule } from './first/first.module';
 import { ProductsModule } from './products/products.module';
 import { CartModule } from './cart/cart.module';
 
@@ -13,6 +14,7 @@ import { CartModule } from './cart/cart.module';
   imports: [
     BrowserModule,
     AppRoutingModule,
+    FirstModule,
     ProductsModule,
     CartModule
   ],