+++ src/app/core/shared/constants.interface.ts
@@ -1,0 +1,4 @@
+export interface Constants {
+  app?: string;
+  ver?: string;
+}