+++ src/app/products/components/product-list/product-list.component.ts
@@ -21,6 +21,6 @@ export class ProductListComponent implements OnInit {
   }
 
   onBuyProduct(product: Product): void {
-    this.cartService.increaseProductCount(product);
+    this.cartService.addProduct(product, 1);
   }
 }