+++ src/app/shared/directives/border.directive.spec.ts
@@ -1,0 +1,8 @@
+import { BorderDirective } from './border.directive';
+
+describe('HostClickDirective', () => {
+  it('should create an instance', () => {
+    const directive = new BorderDirective(null, null);
+    expect(directive).toBeTruthy();
+  });
+});